Add eslint, update packages, code & structure changes
