+++ parse.js
@@ -1,4 +1,4 @@
 require("./src/mongoose").connect();
 
 require("./src/parsers/facebook");
-require("./src/parsers/tutby");
+// require("./src/parsers/tutby");
