+++ src/parsers/facebook/selectors.js
@@ -1,0 +1,10 @@
+/* eslint-disable no-undef */
+module.exports = {
+  eventLink: "div.clearfix > div > div > div > div > div > a",
+  progressBar: "span[role=\"progressbar\"]:not([value=\"jewelLoading\"])",
+  tickets: "[data-testid='event_ticket_link']",
+  title: "#seo_h1_tag",
+  description: "[data-testid=event-permalink-details] > span",
+  time: "#event_time_info div[content]",
+  image: "#event_header_primary img.img",
+};
