+++ src/parsers/tutby/config.js
@@ -1,12 +1,12 @@
-const moment = require('moment');
+const moment = require("moment");
 
-const dateRangeFormat = 'YYYY-MM-DD';
-const singleDateFormat = 'YYYY/MM/DD';
+const dateRangeFormat = "YYYY-MM-DD";
+const singleDateFormat = "YYYY/MM/DD";
 
 module.exports = {
   eventsPageUrl: (startDate, endDate = null) => {
     if (!startDate) {
-      return 'https://afisha.tut.by/free-events'
+      return "https://afisha.tut.by/free-events";
     }
 
     if (!endDate) {
@@ -14,8 +14,8 @@ module.exports = {
       return `https://afisha.tut.by/day/free-events/${formattedStartDate}`;
     }
 
-    const formattedStartDate = moment(startDate).format(dateFormat);
-    const formattedEndDate = moment(startDate).format(dateFormat);
+    const formattedStartDate = moment(startDate).format(dateRangeFormat);
+    const formattedEndDate = moment(startDate).format(dateRangeFormat);
     return `https://afisha.tut.by/day/free-events/${formattedStartDate}/${formattedEndDate}`;
   },
 };
