+++ src/parsers/tutby/selectors.js
@@ -5,23 +5,23 @@ module.exports = {
     today: ".daterangepicker.dropdown-menu li:nth-child(1)",
   },
   eventsTable: {
-    table: '#schedule-table',
-    tableRow: '.b-afisha-event',
-    eventList: '.event-list-i',
-    eventLink: '.item-header-i a.header__link',
+    table: "#schedule-table",
+    tableRow: ".b-afisha-event",
+    eventList: ".event-list-i",
+    eventLink: ".item-header-i a.header__link",
   },
   eventPage: {
-    container: '.post_wrapper',
-    title: '#event-name',
-    image: '.main_image',
-    categories: '.sub_title.tag-place a',
+    container: ".post_wrapper",
+    title: "#event-name",
+    image: ".main_image",
+    categories: ".sub_title.tag-place a",
 
-    place: 'a.b-event_place span',
-    placeLink: 'a.b-event_place',
-    address: 'a.b-event_address',
+    place: "a.b-event_place span",
+    placeLink: "a.b-event_place",
+    address: "a.b-event_address",
     
-    description: '#event-description',
-    start: '.b-event__tickets time[itemprop=startDate]',
-    end: '.b-event__tickets time[itemprop=endDate]',
+    description: "#event-description",
+    start: ".b-event__tickets time[itemprop=startDate]",
+    end: ".b-event__tickets time[itemprop=endDate]",
   }
 };
