--- src/parsers/tutby/utils.js
@@ -1,1 +1,0 @@
-const selectors = require('./selectors')
