+++ src/shared/constants.js
@@ -1,0 +1,12 @@
+/* eslint-disable no-undef */
+module.exports = {
+  browserOptions: {
+    args : [
+      "--no-sandbox",
+      "--disable-setuid-sandbox"
+    ],
+    headless: false,
+    timeout: 60000
+  },
+  dateFormat: "DD.MM.YY, HH:mm",
+};
