+++ src/utils/analyze.js
@@ -1,6 +1,42 @@
-const {
-  eventsClassifier
-} = require("./../constants");
+const eventsClassifier = {
+  paid: {
+    tag: "paid",
+    markers: [
+      "билет",
+      "BYN",
+      "платный",
+      "купить",
+      "стоимость",
+      "пригласительный",
+      "руб.",
+      "білет",
+      "платны",
+      "купіць",
+      "кошт",
+      "запрашальны"
+    ]
+  },
+  free: {
+    tag: "free",
+    markers: [
+      "бесплатный вход",
+      "бесплатно",
+      "свободный",
+      "безвозмездный",
+      "халявный",
+      "регистрация",
+      "вход по регистрации",
+      "бясплатны ўваход",
+      "бясплатна",
+      "свабодны",
+      "вольны",
+      "бязвыплатны",
+      "халяўны",
+      "рэгістрацыя",
+      "уваход па рэгістрацыі"
+    ]
+  }
+};
 
 const natural = require("natural");
 const porterStemmer = natural.PorterStemmerRu;
